requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.6
apache-airflow==2.7.0
sqlalchemy==1.4.46